+++ Source/CDR.DataHolder.API.Infrastructure/Models/AuthorisationPolicy.cs
@@ -5,6 +5,7 @@
         GetCustomersApi,
         GetAccountsApi,
         GetTransactionsApi,
+        GetMetricsApi,
         DynamicClientRegistration
     }
 }