+++ Source/CDR.DataHolder.Repository/Infrastructure/MappingProfile.cs
@@ -1,7 +1,6 @@
 ﻿using AutoMapper;
 using CDR.DataHolder.Repository.Entities;
 using DomainEntities = CDR.DataHolder.Domain.Entities;
-using DomainValueObjects = CDR.DataHolder.Domain.ValueObjects;
 namespace CDR.DataHolder.Repository.Infrastructure
 {
 	public class MappingProfile : Profile
@@ -36,16 +35,6 @@ namespace CDR.DataHolder.Repository.Infrastructure
 				.IncludeMembers(source => source.Organisation)
 				.ReverseMap();
 
-			CreateMap<Brand, DomainEntities.Brand>()
-				.ForMember(dest => dest.LegalEntity, source => source.MapFrom(source => source.LegalEntity))
-				.ReverseMap();
-
-			CreateMap<LegalEntity, DomainEntities.LegalEntity>()
-				.ReverseMap();
-
-			CreateMap<SoftwareProduct, DomainEntities.SoftwareProduct>()
-				.ForMember(dest => dest.Brand, source => source.MapFrom(source => source.Brand))
-				.ReverseMap();
 		}
 	}
 }